update more examples & cleanups
