+++ packages/fluentui/docs/src/examples/components/Button/States/ButtonExampleLoading.shorthand.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Button } from '@fluentui/react-northstar';
+import { Button } from '@fluentui/react-button';
 import { useBooleanKnob } from '@fluentui/docs-components';
 
 const ButtonExampleLoading = () => {
