+++ packages/fluentui/docs/src/examples/components/Button/Usage/ButtonExampleOverflow.shorthand.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Button } from '@fluentui/react-northstar';
+import { Button } from '@fluentui/react-button';
 
 const ButtonExampleOverflow = () => <Button content="See how this very long text shows up in the button" />;
