+++ packages/fluentui/docs/src/examples/components/Button/Usage/ButtonExampleWithTooltip.shorthand.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { Button, Tooltip } from '@fluentui/react-northstar';
+import { Tooltip } from '@fluentui/react-northstar';
+import { Button } from '@fluentui/react-button';
 
 const ButtonExampleWithTooltip = () => (
   <Tooltip trigger={<Button content="Button with tooltip" />} content="This is an actionable element." />
