+++ packages/fluentui/docs/src/examples/components/Button/Variations/ButtonExampleCircular.shorthand.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { Button, Flex } from '@fluentui/react-northstar';
+import { Flex } from '@fluentui/react-northstar';
+import { Button } from '@fluentui/react-button';
 import { EmojiIcon, BroadcastIcon } from '@fluentui/react-icons-northstar';
 
 const ButtonExampleCircular = () => (
