+++ packages/fluentui/docs/src/examples/components/Button/Variations/ButtonExampleFluid.shorthand.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
-import { Button } from '@fluentui/react-northstar';
+import { Button } from '@fluentui/react-button';
 
-const ButtonExampleFluid = () => <Button fluid content="Fits to Container" />;
+const ButtonExampleFluid = () => <Button block content="Fits to Container" />;
 
 export default ButtonExampleFluid;
